// src/app/general-module/componentes-comunes/interfaces/DenunciaGrabada.ts
export interface DenunciaGrabada {
    auditor?: string;
    correlativo?: string;
    departamento?: number;
    direccionEstDenunciado?: string;
    direccionFiscalDenunciado?: string;
    email?: string;
    establecimientoDenunciado?: string;
    estado?: string;
    fechaAsignacion?: Date;
    fechaAsignacionJs?: Date;
    fechaCargaPlan?: Date;
    fechaCompra?: string;
    fechaGrabacion?: Date;
    fechaGrabacionStr?: string;
    fechaRechazo?: Date;
    formaPago?: string;
    idInsumo?: number;
    idMotivo?: number;
    justificacion?: string;
    latitud?: string;
    linea?: number;
    longitud?: string;
    municipio?: number;
    nitAp?: string;
    nitAuditorRechazo?: string;
    nitDenunciado?: string;
    nitDenunciante?: string;
    nitJs?: string;
    nitSn?: string;
    nombreDenunciado?: string;
    nombreDenunciante?: string;
    nombrePlan?: string;
    observaciones?: string;
    otroMotivo?: string;
    productoServicio?: string;
    razonRechazo?: string;
    region?: number;
    telefonoDenunciado?: string;
    telefonoDenunciante?: string;
    urlImagen?: string;
    valorCompraServicio?: number;
    pFechaFin?: string;
    pFechaInicio?: string;
    idProceso?: number;
    idZona?: number;
    idGrupoHabitacional?: number;
    nombreColonia?: string;
    idVialidad?: number;
    idSubConceptoDenuncia?: number;
}